requirements.txt
typer[all]>=0.9.0
lkml>=1.3.1
pyyaml>=6.0.1
rich>=13.7.1
requests>=2.32.4
pydoc-markdown>=4.8.2
